refactor: move repeating code to common controller
